--- domain/src/main/java/com/aireview/review/config/JpaConfiguration.java
@@ -1,9 +1,0 @@
-package com.aireview.review.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaConfig {
-}
